restructured as index.js
