+++ server/server.js
@@ -31,7 +31,7 @@ const signup = require("./routes/userAuth/signup");
 const getToken = require("./routes/userAuth/getToken");
 const search = require("./routes/products/searchProducts");
 const address = require("./routes/order/address");
-const cart = require("./routes/cart/cart.routes");
+const cart = require("./routes/cart");
 const payment = require("./routes/order/payment");
 const orders = require("./routes/order/orders");
 // const checkout = require("./routes/checkout");
